added migrations and load correct address
