+++ OpenVPNGateMonitor.Models/Helpers/OpenVpnSettings.cs
@@ -8,6 +8,7 @@ public class OpenVpnSettings
     public string ServerIp { get; init; } = "213.133.91.43";
     public string ManagementIp { get; init; } = "127.0.0.1";
     public int ManagementPort { get; init; } = 5555;
+    public int UpdateSecond { get; init; } = 120;
     public string CrlPkiPath { get; init; } = "/etc/openvpn/easy-rsa/pki/crl.pem";
     public string CrlOpenvpnPath { get; init; } = "/etc/openvpn/crl.pem";
     public string StatusFilePath { get; init; } = "/var/log/openvpn-status.log";
