--- OpenVPNGateMonitor/Services/Api/Interfaces/IOpenVpnServerService.cs
@@ -1,7 +1,0 @@
-﻿namespace OpenVPNGateMonitor.Services.Api.Interfaces;
-
-public interface IOpenVpnServerService
-{
-    Task SaveConnectedClientsAsync(CancellationToken cancellationToken);
-    Task SaveOpenVpnServerStatusLogAsync(CancellationToken cancellationToken);
-}
