+++ OpenVPNGateMonitor/Services/BackgroundServices/Interfaces/IOpenVpnServerService.cs
@@ -1,0 +1,7 @@
+﻿namespace OpenVPNGateMonitor.Services.BackgroundServices.Interfaces;
+
+public interface IOpenVpnServerService
+{
+    Task SaveConnectedClientsAsync(CancellationToken cancellationToken);
+    Task SaveOpenVpnServerStatusLogAsync(CancellationToken cancellationToken);
+}
