+++ OpenVPNGateMonitor/Services/BackgroundServices/OpenVpnBackgroundService.cs
@@ -1,5 +1,5 @@
-﻿using OpenVPNGateMonitor.Services.BackgroundServices.Interfaces;
-using OpenVPNGateMonitor.Services.OpenVpnManagementInterfaces.Interfaces;
+﻿using OpenVPNGateMonitor.Models.Helpers;
+using OpenVPNGateMonitor.Services.BackgroundServices.Interfaces;
 
 namespace OpenVPNGateMonitor.Services.BackgroundServices;
 
@@ -7,72 +7,49 @@ public class OpenVpnBackgroundService : BackgroundService, IOpenVpnBackgroundSer
 {
     private readonly ILogger<OpenVpnBackgroundService> _logger;
     private readonly IServiceProvider _serviceProvider;
-    private const int Seconds = 120;
+    private readonly int _seconds;
 
-    public OpenVpnBackgroundService(ILogger<OpenVpnBackgroundService> logger, IServiceProvider serviceProvider)
+    public OpenVpnBackgroundService(ILogger<OpenVpnBackgroundService> logger, 
+        IServiceProvider serviceProvider, IConfiguration configuration)
     {
         _logger = logger;
         _serviceProvider = serviceProvider;
-    }
-
-    protected override async Task ExecuteAsync(CancellationToken stoppingToken)
-    {
-        _logger.LogInformation("OpenVPN Background Service is starting.");
-
-        while (!stoppingToken.IsCancellationRequested)
+        var openVpnSection = configuration.GetSection("OpenVpn");
+        if (!openVpnSection.Exists())
         {
-            using (var scope = _serviceProvider.CreateScope())
-            {
-                var openVpnStateService = scope.ServiceProvider.GetRequiredService<IOpenVpnStateService>();
-                try
-                {
-                    var state = await openVpnStateService.GetStateAsync(stoppingToken);
-                    _logger.LogInformation($"State: {state.RemoteIp} {state.Success} {state.Connected} {state.LocalIp} ");
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error occurred while parsing OpenVPN state.");
-                }
-                var openVpnSummaryStatService = scope.ServiceProvider.GetRequiredService<IOpenVpnSummaryStatService>();
-                try
-                {
-                    var summaryStats = await openVpnSummaryStatService.GetSummaryStatsAsync(stoppingToken);
-                    _logger.LogInformation($"Summary stats: {summaryStats.BytesIn} {summaryStats.BytesOut} {summaryStats.ClientsCount} ");
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error occurred while parsing OpenVPN summary stats.");
-                }
-                var openVpnClientService = scope.ServiceProvider.GetRequiredService<IOpenVpnClientService>();
-                try
-                {
-                    var clients= await openVpnClientService.GetClientsAsync(stoppingToken);
-                    foreach (var client in clients)
-                    {
-                        _logger.LogInformation($"Clients: {client.CommonName} {client.RemoteIp} {client.LocalIp} {client.BytesReceived} {client.BytesSent} {client.Country}");
-                        _logger.LogInformation($"Clients: {client.Country} {client.Region} {client.City} {client.Latitude} {client.Longitude}");
-                    }
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error occurred while parsing OpenVPN clients.");
-                }
-                
-                var openVpnVersionService = scope.ServiceProvider.GetRequiredService<IOpenVpnVersionService>();
-                try
-                {
-                    var version= await openVpnVersionService.GetVersionAsync(stoppingToken);
-                    _logger.LogInformation($"Version: {version}");
-                }
-                catch (Exception ex)
-                {
-                    _logger.LogError(ex, "Error occurred while parsing OpenVPN version.");
-                }
-            }
-
-            await Task.Delay(TimeSpan.FromSeconds(Seconds), stoppingToken);
+            throw new InvalidOperationException("OpenVpn section is missing in the configuration.");
         }
 
-        _logger.LogInformation("OpenVPN Background Service is stopping.");
+        _seconds = openVpnSection.Get<OpenVpnSettings>()!.UpdateSecond;
+        if (_seconds <= 0)
+        {
+            throw new InvalidOperationException("Failed to load OpenVpnSettings UpdateSecond from configuration.");
+        }
+    }
+
+    protected override async Task ExecuteAsync(CancellationToken cancellationToken)
+    {
+        try
+        {
+            _logger.LogInformation("OpenVPN Background Service is starting.");
+
+            while (!cancellationToken.IsCancellationRequested)
+            {
+                using (var scope = _serviceProvider.CreateScope())
+                {
+                    var openVpnServerService = scope.ServiceProvider.GetRequiredService<IOpenVpnServerService>();
+                    await openVpnServerService.SaveOpenVpnServerStatusLogAsync(cancellationToken);
+                    await openVpnServerService.SaveConnectedClientsAsync(cancellationToken);
+                }
+
+                await Task.Delay(TimeSpan.FromSeconds(_seconds), cancellationToken);
+            }
+
+            _logger.LogInformation("OpenVPN Background Service is stopping.");
+        }
+        catch(Exception ex)
+        {
+            _logger.LogError(ex, $"An error occured while running the background service. Error: {ex.Message}");
+        }
     }
 }
