--- OpenVPNGateMonitor/Services/Interfaces/IOpenVpnParserService.cs
@@ -1,6 +1,0 @@
-﻿namespace OpenVPNGateMonitor.Services.Interfaces;
-
-public interface IOpenVpnParserService
-{
-    Task ParseAndSaveAsync();
-}
