+++ OpenVPNGateMonitor/Services/OpenVpnManagementInterfaces/GeoIpService.cs
@@ -1,4 +1,6 @@
-﻿using MaxMind.GeoIP2;
+﻿using System.Net;
+using MaxMind.GeoIP2;
+using MaxMind.GeoIP2.Exceptions;
 using MaxMind.GeoIP2.Responses;
 using OpenVPNGateMonitor.Models.Helpers;
 using OpenVPNGateMonitor.Models.Helpers.OpenVpnManagementInterfaces;
@@ -8,10 +10,12 @@ namespace OpenVPNGateMonitor.Services.OpenVpnManagementInterfaces;
 
 public class GeoIpService : IGeoIpService
 {
+    private readonly ILogger<IGeoIpService> _logger;
     private readonly DatabaseReader? _geoIpReader;
 
-    public GeoIpService(IConfiguration configuration)
+    public GeoIpService(ILogger<IGeoIpService> logger, IConfiguration configuration)
     {
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         var openVpnSettings = configuration.GetSection("OpenVpn").Get<OpenVpnSettings>() 
                               ?? throw new InvalidOperationException("OpenVpn configuration section is missing.");
 
@@ -24,24 +28,61 @@ public class GeoIpService : IGeoIpService
     
     public OpenVpnGeoInfo? GetGeoInfo(string ip)
     {
-        if (_geoIpReader == null) return null;
-
         try
         {
-            CityResponse cityResponse = _geoIpReader.City(ip);
+            if (string.IsNullOrWhiteSpace(ip))
+                return null;
+
+            var ipAddress = IPAddress.Parse(ip);
+            
+            if (IsPrivateIp(ipAddress))
+            {
+                return new OpenVpnGeoInfo
+                {
+                    Country = "Internet",
+                    Region = "RFC1918",
+                    City = "RFC1918",
+                    Latitude = 0,
+                    Longitude = 0
+                };//85.172.77.167
+            }
+
+            if (ipAddress.IsIPv6LinkLocal || ipAddress.IsIPv6Multicast)
+                return null;
+
+            CityResponse cityResponse = _geoIpReader!.City(ip);
+
             return new OpenVpnGeoInfo
             {
-                Country = cityResponse.Country.IsoCode,
-                Region = cityResponse.MostSpecificSubdivision.IsoCode,
-                City = cityResponse.City.Name,
-                Latitude = cityResponse.Location.Latitude,
-                Longitude = cityResponse.Location.Longitude
+                Country = cityResponse.RegisteredCountry?.IsoCode ?? cityResponse.Country?.IsoCode,
+                Region = cityResponse.MostSpecificSubdivision?.IsoCode 
+                         ?? cityResponse.Subdivisions.LastOrDefault()?.IsoCode
+                         ?? cityResponse.RegisteredCountry?.IsoCode,
+                City = cityResponse.City?.Name 
+                       ?? cityResponse.Location?.TimeZone 
+                       ?? cityResponse.RegisteredCountry?.Name,
+                Latitude = cityResponse.Location?.Latitude,
+                Longitude = cityResponse.Location?.Longitude
             };
+        }
+        catch (AddressNotFoundException ex)
+        {
+            _logger.LogWarning($"GeoIP not found for {ip}: {ex.Message}");
+            return null;
         }
         catch (Exception ex)
         {
-            Console.WriteLine($"GeoIP Error: {ex.Message}");
+            _logger.LogError(ex, $"Error getting GeoIP for {ip}");
             return null;
         }
+    }
+
+    private bool IsPrivateIp(IPAddress ip)
+    {
+        byte[] bytes = ip.GetAddressBytes();
+        return (bytes[0] == 10) || // 10.0.0.0/8
+               (bytes[0] == 172 && bytes[1] >= 16 && bytes[1] <= 31) || // 172.16.0.0/12
+               (bytes[0] == 192 && bytes[1] == 168) || // 192.168.0.0/16
+               IPAddress.IsLoopback(ip);
     }
 }
